+++ internal/checks/promql_series.go
@@ -3,10 +3,15 @@ package checks
 import (
 	"context"
 	"fmt"
+	"time"
 
+	"github.com/rs/zerolog/log"
+
+	"github.com/cloudflare/pint/internal/output"
 	"github.com/cloudflare/pint/internal/parser"
 	"github.com/cloudflare/pint/internal/promapi"
 
+	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
 	promParser "github.com/prometheus/prometheus/promql/parser"
 )
@@ -38,9 +43,17 @@ func (c SeriesCheck) Check(ctx context.Context, rule parser.Rule) (problems []Pr
 		return
 	}
 
-	done := map[string]bool{}
+	rangeStart := time.Now().Add(time.Hour * 24 * -7)
+	rangeStep := time.Minute * 5
 
+	done := map[string]bool{}
 	for _, selector := range getSelectors(expr.Query) {
+		if _, ok := done[selector.String()]; ok {
+			continue
+		}
+
+		done[selector.String()] = true
+
 		bareSelector := stripLabels(selector)
 		c1 := fmt.Sprintf("disable %s(%s)", SeriesCheckName, selector.String())
 		c2 := fmt.Sprintf("disable %s(%s)", SeriesCheckName, bareSelector.String())
@@ -48,29 +61,219 @@ func (c SeriesCheck) Check(ctx context.Context, rule parser.Rule) (problems []Pr
 			done[selector.String()] = true
 			continue
 		}
-		if _, ok := done[selector.String()]; ok {
+
+		metricName := selector.Name
+		if metricName == "" {
+			for _, lm := range selector.LabelMatchers {
+				if lm.Name == labels.MetricName && lm.Type == labels.MatchEqual {
+					metricName = lm.Value
+					break
+				}
+			}
+		}
+
+		labelNames := []string{}
+		for _, lm := range selector.LabelMatchers {
+			if lm.Name != labels.MetricName {
+				labelNames = append(labelNames, lm.Name)
+			}
+		}
+
+		// 1. If foo{bar, baz} is there -> GOOD
+		log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Msg("Checking if selector returns anything")
+		count, _, err := c.instantSeriesCount(ctx, fmt.Sprintf("count(%s)", selector.String()))
+		if err != nil {
+			problems = append(problems, c.queryProblem(err, selector.String(), expr))
 			continue
 		}
-		problems = append(problems, c.countSeries(ctx, expr, selector)...)
-		done[selector.String()] = true
+		if count > 0 {
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Msg("Found series, skipping further checks")
+			continue
+		}
+
+		// 2. If foo was NEVER there -> BUG
+		log.Debug().Str("check", c.Reporter()).Stringer("selector", &bareSelector).Msg("Checking if base metric has historical series")
+		trs, err := c.serieTimeRanges(ctx, fmt.Sprintf("count(%s)", bareSelector.String()), rangeStart, rangeStep)
+		if err != nil {
+			problems = append(problems, c.queryProblem(err, bareSelector.String(), expr))
+			continue
+		}
+		if len(trs.ranges) == 0 {
+			problems = append(problems, Problem{
+				Fragment: bareSelector.String(),
+				Lines:    expr.Lines(),
+				Reporter: c.Reporter(),
+				Text: fmt.Sprintf("%s didn't have any series for %q metric in the last %s",
+					promText(c.prom.Name(), trs.uri), bareSelector.String(), trs.sinceDesc()),
+				Severity: Bug,
+			})
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &bareSelector).Msg("No historical series for base metric")
+			continue
+		}
+
+		highChurnLabels := []string{}
+
+		// 3. If foo is ALWAYS/SOMETIMES there BUT {bar OR baz} is NEVER there -> BUG
+		for _, name := range labelNames {
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Str("label", name).Msg("Checking if base metric has historical series with required label")
+			trsLabelCount, err := c.serieTimeRanges(ctx, fmt.Sprintf("count(%s) by (%s)", bareSelector.String(), name), rangeStart, rangeStep)
+			if err != nil {
+				problems = append(problems, c.queryProblem(err, bareSelector.String(), expr))
+				continue
+			}
+
+			labelRanges := trsLabelCount.withLabelName(name)
+			if len(labelRanges) == 0 {
+				problems = append(problems, Problem{
+					Fragment: selector.String(),
+					Lines:    expr.Lines(),
+					Reporter: c.Reporter(),
+					Text: fmt.Sprintf(
+						"%s has %q metric but there are no series with %q label in the last %s",
+						promText(c.prom.Name(), trsLabelCount.uri), bareSelector.String(), name, trsLabelCount.sinceDesc()),
+					Severity: Bug,
+				})
+				log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Str("label", name).Msg("No historical series with label used for the query")
+			}
+
+			if len(trsLabelCount.labelValues(name)) == len(trsLabelCount.ranges) && trsLabelCount.avgLife() < (trsLabelCount.duration()/2) {
+				highChurnLabels = append(highChurnLabels, name)
+			}
+		}
+		if len(problems) > 0 {
+			continue
+		}
+
+		// 4. If foo was ALWAYS there but it's NO LONGER there -> BUG
+		if len(trs.ranges) == 1 &&
+			!trs.oldest().After(rangeStart.Add(rangeStep)) &&
+			trs.newest().Before(time.Now().Add(rangeStep*-1)) {
+			problems = append(problems, Problem{
+				Fragment: bareSelector.String(),
+				Lines:    expr.Lines(),
+				Reporter: c.Reporter(),
+				Text: fmt.Sprintf(
+					"%s doesn't currently have %q, it was last present %s ago",
+					promText(c.prom.Name(), trs.uri), bareSelector.String(), output.HumanizeDuration(time.Since(trs.newest()))),
+				Severity: Bug,
+			})
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &bareSelector).Msg("Series disappeared from prometheus ")
+			continue
+		}
+
+		for _, lm := range selector.LabelMatchers {
+			if lm.Name == labels.MetricName {
+				continue
+			}
+			if lm.Type != labels.MatchEqual && lm.Type != labels.MatchRegexp {
+				continue
+			}
+			labelSelector := promParser.VectorSelector{
+				Name:          metricName,
+				LabelMatchers: []*labels.Matcher{lm},
+			}
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &labelSelector).Stringer("matcher", lm).Msg("Checking if there are historical series matching filter")
+
+			trsLabel, err := c.serieTimeRanges(ctx, fmt.Sprintf("count(%s)", labelSelector.String()), rangeStart, rangeStep)
+			if err != nil {
+				problems = append(problems, c.queryProblem(err, bareSelector.String(), expr))
+				continue
+			}
+
+			// 5. If foo is ALWAYS/SOMETIMES there BUT {bar OR baz} value is NEVER there -> BUG
+			if len(trsLabel.ranges) == 0 {
+				text := fmt.Sprintf(
+					"%s has %q metric but there are no series matching {%s} in the last %s",
+					promText(c.prom.Name(), trsLabel.uri), bareSelector.String(), lm.String(), trsLabel.sinceDesc())
+				var s Severity = Bug
+				for _, name := range highChurnLabels {
+					if lm.Name == name {
+						s = Warning
+						text += fmt.Sprintf(", %q looks like a high churn label", name)
+						break
+					}
+				}
+
+				problems = append(problems, Problem{
+					Fragment: selector.String(),
+					Lines:    expr.Lines(),
+					Reporter: c.Reporter(),
+					Text:     text,
+					Severity: s,
+				})
+				log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Stringer("matcher", lm).Msg("No historical series matching filter used in the query")
+				continue
+			}
+
+			// 6. If foo is ALWAYS/SOMETIMES there AND {bar OR baz} used to be there ALWAYS BUT it's NO LONGER there -> BUG
+			if len(trsLabel.ranges) == 1 &&
+				!trsLabel.oldest().After(rangeStart.Add(rangeStep)) &&
+				trsLabel.newest().Before(time.Now().Add(rangeStep*-1)) {
+				problems = append(problems, Problem{
+					Fragment: bareSelector.String(),
+					Lines:    expr.Lines(),
+					Reporter: c.Reporter(),
+					Text: fmt.Sprintf(
+						"%s has %q metric but doesn't currently have series matching {%s}, such series was last present %s ago",
+						promText(c.prom.Name(), trs.uri), bareSelector.String(), lm.String(), output.HumanizeDuration(time.Since(trs.newest()))),
+					Severity: Bug,
+				})
+				log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Stringer("matcher", lm).Msg("Series matching filter disappeared from prometheus ")
+				continue
+			}
+
+			// 7. if foo is ALWAYS/SOMETIMES there BUT {bar OR baz} value is SOMETIMES there -> WARN
+			if len(trsLabel.ranges) > 1 {
+				problems = append(problems, Problem{
+					Fragment: selector.String(),
+					Lines:    expr.Lines(),
+					Reporter: c.Reporter(),
+					Text: fmt.Sprintf(
+						"metric %q with label %s is only sometimes present on %s with average life span of %s",
+						bareSelector.String(), lm.String(), promText(c.prom.Name(), trs.uri),
+						output.HumanizeDuration(trs.avgLife())),
+					Severity: Warning,
+				})
+				log.Debug().Str("check", c.Reporter()).Stringer("selector", &selector).Stringer("matcher", lm).Msg("Series matching filter are only sometimes present")
+			}
+		}
+		if len(problems) > 0 {
+			continue
+		}
+
+		// 8. If foo is SOMETIMES there -> WARN
+		if len(trs.ranges) > 1 {
+			problems = append(problems, Problem{
+				Fragment: selector.String(),
+				Lines:    expr.Lines(),
+				Reporter: c.Reporter(),
+				Text: fmt.Sprintf(
+					"metric %q is only sometimes present on %s with average life span of %s in the last %s",
+					bareSelector.String(), promText(c.prom.Name(), trs.uri), output.HumanizeDuration(trs.avgLife()), trs.sinceDesc()),
+				Severity: Warning,
+			})
+			log.Debug().Str("check", c.Reporter()).Stringer("selector", &bareSelector).Msg("Metric only sometimes present")
+		}
 	}
 
 	return
 }
 
-func (c SeriesCheck) countSeries(ctx context.Context, expr parser.PromQLExpr, selector promParser.VectorSelector) (problems []Problem) {
-	q := fmt.Sprintf("count(%s)", selector.String())
-	qr, err := c.prom.Query(ctx, q)
+func (c SeriesCheck) queryProblem(err error, selector string, expr parser.PromQLExpr) Problem {
+	text, severity := textAndSeverityFromError(err, c.Reporter(), c.prom.Name(), Bug)
+	return Problem{
+		Fragment: selector,
+		Lines:    expr.Lines(),
+		Reporter: c.Reporter(),
+		Text:     text,
+		Severity: severity,
+	}
+}
+
+func (c SeriesCheck) instantSeriesCount(ctx context.Context, query string) (int, string, error) {
+	qr, err := c.prom.Query(ctx, query)
 	if err != nil {
-		text, severity := textAndSeverityFromError(err, c.Reporter(), c.prom.Name(), Bug)
-		problems = append(problems, Problem{
-			Fragment: selector.String(),
-			Lines:    expr.Lines(),
-			Reporter: c.Reporter(),
-			Text:     text,
-			Severity: severity,
-		})
-		return
+		return 0, "", err
 	}
 
 	var series int
@@ -78,30 +281,48 @@ func (c SeriesCheck) countSeries(ctx context.Context, expr parser.PromQLExpr, se
 		series += int(s.Value)
 	}
 
-	if series == 0 {
-		if len(selector.LabelMatchers) > 1 {
-			// retry selector with only __name__ label
-			s := stripLabels(selector)
-			p := c.countSeries(ctx, expr, s)
-			// if we have zero series without any label selector then the whole
-			// series is missing, but if we have some then report missing series
-			// with labels
-			if len(p) > 0 {
-				problems = append(problems, p...)
-				return
-			}
-		}
-		problems = append(problems, Problem{
-			Fragment: selector.String(),
-			Lines:    expr.Lines(),
-			Reporter: c.Reporter(),
-			Text:     fmt.Sprintf("query using %q on %s completed without any results for %s", c.prom.Name(), qr.URI, selector.String()),
-			Severity: Bug,
-		})
-		return
+	return series, qr.URI, nil
+}
+
+func (c SeriesCheck) serieTimeRanges(ctx context.Context, query string, from time.Time, step time.Duration) (tr *timeRanges, err error) {
+	now := time.Now()
+	qr, err := c.prom.RangeQuery(ctx, query, from, now, step)
+	if err != nil {
+		return nil, err
 	}
 
-	return nil
+	tr = &timeRanges{
+		uri:   qr.URI,
+		from:  from,
+		until: now,
+		step:  step,
+	}
+	var ts time.Time
+	for _, s := range qr.Samples {
+		for _, v := range s.Values {
+			ts = v.Timestamp.Time()
+
+			var found bool
+			for i := range tr.ranges {
+				if tr.ranges[i].labels.Equal(model.LabelSet(s.Metric)) &&
+					!ts.Before(tr.ranges[i].start) &&
+					!ts.After(tr.ranges[i].end) {
+					tr.ranges[i].end = ts.Add(step)
+					found = true
+					break
+				}
+			}
+			if !found {
+				tr.ranges = append(tr.ranges, timeRange{
+					labels: model.LabelSet(s.Metric),
+					start:  ts,
+					end:    ts.Add(step),
+				})
+			}
+		}
+	}
+
+	return tr, nil
 }
 
 func getSelectors(n *parser.PromQLNode) (selectors []promParser.VectorSelector) {
@@ -132,4 +353,84 @@ func stripLabels(selector promParser.VectorSelector) promParser.VectorSelector {
 		}
 	}
 	return s
+}
+
+type timeRange struct {
+	labels model.LabelSet
+	start  time.Time
+	end    time.Time
+}
+
+type timeRanges struct {
+	uri    string
+	from   time.Time
+	until  time.Time
+	step   time.Duration
+	ranges []timeRange
+}
+
+func (tr timeRanges) withLabelName(name string) (r []timeRange) {
+	for _, s := range tr.ranges {
+		for k := range s.labels {
+			if k == model.LabelName(name) {
+				r = append(r, s)
+			}
+		}
+	}
+	return
+}
+
+func (tr timeRanges) labelValues(name string) (vals []string) {
+	vm := map[string]struct{}{}
+	for _, s := range tr.ranges {
+		for k, v := range s.labels {
+			if k == model.LabelName(name) {
+				vm[string(v)] = struct{}{}
+			}
+		}
+	}
+	for v := range vm {
+		vals = append(vals, v)
+	}
+	return
+}
+
+func (tr timeRanges) duration() (d time.Duration) {
+	return tr.until.Sub(tr.from)
+}
+
+func (tr timeRanges) avgLife() (d time.Duration) {
+	for _, r := range tr.ranges {
+		d += r.end.Sub(r.start)
+	}
+	if len(tr.ranges) == 0 {
+		return time.Duration(0)
+	}
+	return time.Second * time.Duration(int(d.Seconds())/len(tr.ranges))
+}
+
+func (tr timeRanges) oldest() (ts time.Time) {
+	for _, r := range tr.ranges {
+		if ts.IsZero() || r.start.Before(ts) {
+			ts = r.start
+		}
+	}
+	return
+}
+
+func (tr timeRanges) newest() (ts time.Time) {
+	for _, r := range tr.ranges {
+		if ts.IsZero() || r.end.After(ts) {
+			ts = r.end
+		}
+	}
+	return
+}
+
+func (tr timeRanges) sinceDesc() (s string) {
+	dur := time.Since(tr.from)
+	if dur > time.Hour*24 {
+		return output.HumanizeDuration(dur.Round(time.Hour))
+	}
+	return output.HumanizeDuration(dur.Round(time.Minute))
 }