+++ docs/changelog.md
@@ -5,7 +5,8 @@
 ### Changed
 
 - Always print the number of detected problems when running `pint lint`.
-- Raise `promql/series` severity from `Warning` to `Bug`.
+- `promql/series` check was refactored and will now detect a range of
+  problems. See [promql/series](checks/promql/series.md) for details.
 
 ## v0.14.0